Refactored computergroups, policies, extended policies
